# docs/SJewls_Admin_API_OpenAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SJewls Admin & Staff Portal API",
    "description": "API endpoints for the SJewls Admin Web Portal \u2014 Staff Authentication, Password Recovery, Staff User Management, Customer Management with Full Unmasked NIC, and Branch Statistics.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5230",
      "description": "Current Host"
    },
    {
      "url": "http://localhost:5230",
      "description": "Local Server"
    }
  ],
  "paths": {
...
    "/api/v1/admin/users": {
      "get": {
        "tags": [
          "Admin Users"
        ],
        "summary": "List Staff Users",
        "description": "Retrieves a list of staff accounts. Super Admins view all staff; Branch Admins view staff within their assigned branch. Password hashes are never returned.",
        "operationId": "AdminGetUsers",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StaffUserDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Admin Users"
        ],
        "summary": "Create Staff User",
        "description": "Allows Super Admin to provision a new staff or admin user, assigning their role and branch. Validates email, phone, and password policy, and prevents duplicates.",
        "operationId": "AdminCreateUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStaffUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffUserDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "409": {
            "description": "Conflict"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/v1/admin/users/{id}/status": {
      "patch": {
        "tags": [
          "Admin Users"
        ],
        "summary": "Activate or Deactivate Staff User",
        "description": "Allows authorized administrators to activate or deactivate a staff member. Super Admin can manage all staff and admins (preventing deactivation of the last Super Admin). Staff cannot deactivate themselves. Sessions are immediately invalidated upon deactivation.",
        "operationId": "AdminUpdateUserStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStaffStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffUserDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateStaffUserRequest": {
        "required": [
          "email",
          "fullName",
          "password",
          "phoneNumber",
          "role"
        ],
        "type": "object",
        "properties": {
          "fullName": {
            "maxLength": 150,
            "minLength": 2,
            "type": "string"
          },
          "email": {
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "phoneNumber": {
            "minLength": 1,
            "type": "string"
          },
          "password": {
            "maxLength": 100,
            "minLength": 8,
            "type": "string"
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "minLength": 1,
            "type": "string"
          },
          "branchId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "StaffBranchDto": {
        "type": "object",
        "properties": {
          "branchId": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "StaffUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "assignedBranches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StaffBranchDto"
            },
            "nullable": true
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deactivatedAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deactivationReason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateStaffStatusRequest": {
        "required": [
          "isActive"
        ],
        "type": "object",
        "properties": {
          "isActive": {
            "type": "boolean"
          },
          "reason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter your JWT Bearer token (e.g., Bearer eyJhbGci...)",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}